+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.0 (WIP)
 
 * Upgrade to Alchemy v1.4.0
+* Add edit template for the `Object` field type
 
 ## 0.2.10 (2024-01-15)