+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0-alpha.2 (WIP)
+
+* Added `.alchemy-badge` class
+
 ## 0.3.0-alpha.1 (2024-02-15)
 
 * Upgrade to Alchemy v1.4.0